Restrcture code base into smaller files
